feat(frontend): prefetch models for new metric form
